requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
altair==5.0.1
anyio==3.7.1
apache-beam==2.48.0
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
autotrain-advanced==0.6.25
bitsandbytes==0.41.1
boto3==1.28.22
botocore==1.31.22
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
cloudpickle==2.2.1
cmake==3.26.4
codecarbon==2.2.3
contourpy==1.1.0
crcmod==1.7
cycler==0.11.0
datasets==2.14.4
diffusers @ git+https://github.com/huggingface/diffusers.git@74d902eb59f873b6156621220937f8e2521dfdc0
dill==0.3.6
dnspython==2.4.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
evaluate==0.3.0
exceptiongroup==1.1.2
fastapi==0.101.1
fastavro==1.8.0
fasteners==0.18
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.39.0
gradio_client==0.4.0
grpcio==1.56.0
h11==0.14.0
hdfs==2.7.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
invisible-watermark==0.2.0
ipadic==1.0.0
Jinja2==3.1.2
jiwer==3.0.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
linkify-it-py==2.0.2
lit==16.0.6
loguru==0.7.0
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
objsize==0.6.1
openai==0.27.8
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
orjson==3.9.2
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
peft==0.4.0
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydot==1.4.2
pydub==0.25.1
pymongo==4.4.1
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
rapidfuzz==2.13.7
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
text-generation==0.6.0
threadpoolctl==3.1.0
tifffile==2023.8.12
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers @ git+https://github.com/huggingface/transformers.git@1982dd3b15867c46e1c20645901b0de469fd935f
triton==2.0.0
trl @ git+https://github.com/huggingface/trl.git@029f961b7ced7ad7daf315620effecfdd057ec2b
typing_extensions==4.7.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.23.2
wandb==0.15.5
websockets==11.0.3
Werkzeug==2.3.6
xformers==0.0.21
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
